reader: Use 64bit integer consistenty for node ID
Problem: When a node's ID becomes even larger such that it can't fit
into a signed 32bit integer, our graph loading fails. While it is less
likely we will have such a huge node ID, it would be ideal to prevent
that from happening at all -- in particular it can be done without
massive code changes: our resource pool data class already uses int64_t
for the "id" field.

Change std::atoi to std::atoll to convert a large number string
into a number without raising an overflow exception within a
hostname querying method.

Change the corresponding hostname querying method signatures to take
in int64_t instead of int and adjust all of the callers.

Add an one-line error message when an error such as the overflow
exception occurs to help quicker triaging.

+++ resource/readers/resource_reader_base.hpp
@@ -102,7 +102,7 @@ public:
      *
      * \return          0 if succeeds or -1 with errno if an error encountered
      */
-    int get_hostname_suffix (const std::string &hn, int &id) const;
+    int get_hostname_suffix (const std::string &hn, int64_t &id) const;
 
     /*! Query the host basename from hostname.
      *
@@ -124,7 +124,7 @@ public:
 protected:
     bool in_allowlist (const std::string &resource);
     int split_hostname (const std::string &hn,
-                        std::string &basename, int &id) const;
+                        std::string &basename, int64_t &id) const;
     std::set<std::string> allowlist;
     std::string m_err_msg = "";
 };
